requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
Django==2.2.3
django-webtest==1.9.7
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
factory-boy==2.12.0
Faker==2.0.0
importlib-metadata==0.18
more-itertools==7.1.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.0
pytest==5.0.1
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
text-unidecode==1.2
waitress==1.3.0
wcwidth==0.1.7
WebOb==1.8.5
WebTest==2.0.33
zipp==0.5.2
